Mailing functionality written and partially integrated
Mailing.java is currently called when the program starts and checks the
last time mailScheduler() was ran from a config file. If it has been
more than a week then it will send an update on what the current
statuses are of events with the due date and percentage completed.